Building an AI-Powered Smart City
The increase in urban population in the 21st century has posed urgent challenges to city governments, including increased traffic congestion, pollution, rising energy demand, inadequate healthcare provision, and substandard education and housing systems. Conventional city systems have not been designed to cope with such pressures, and new methods need to be developed that utilize the emerging technologies. Smart cities can be more efficient, sustainable, and responsive to the needs of citizens on a framework based on Artificial Intelligence (AI), the Internet of Things (IoT), and the combination of these two concepts, referred to as the Artificial Intelligence of Things (AIoT). The idea of a smart city that operates using AI and IoT is not merely a compilation of highly developed technologies, but a unified ecosystem that aims to streamline resources and governance and enhance the overall quality of life for its residents. This paper proposes a model of an AI-driven smart city, where key elements of smart healthcare, transportation, housing, education, energy systems, and public infrastructure are integrated. It will describe the resources required to implement it, provide a schedule for its development, and outline how the Information and Communication Technologies (ICT) framework will enhance operational efficiency and facilitate data sharing among sectors.